Improved documentation and added check to verify if path is a directory (in loop and argument)

+++ remove.sh
@@ -3,12 +3,19 @@ ROOT_PATH=$1
 # If path is not manually set, defaults to current working directory
 if [[ ! $ROOT_PATH ]]; then
     ROOT_PATH=$(pwd)
+else
+    # Check if path exists
+    if [[ ! -d $ROOT_PATH ]]; then
+        echo "Path is not valid"
+        exit 1
+    fi 
 fi
+
 
 for dir in $(find $ROOT_PATH -maxdepth 5 -type d)
 do
     # Removed any folder called "node_modules" and ignores any sub-folders
-    if [[ $(basename $dir) == "node_modules" && $(echo "$dir" | grep -o "node_modules" | wc -l) -lt 2 ]]; then
+    if [[ $(basename $dir) == "node_modules" && $(echo "$dir" | grep -o "node_modules" | wc -l) -lt 2 && -d $dir ]]; then
         echo "Deleting $dir..."
         rm -rf $dir
     fi
